Added support for fault tolerant coroutines.
Fault tolerance is supported through 'monitor' coroutines, so that
a monitor can watch how a coroutine is finished and possibly restart
the coroutine with (a different) generator. If that monitor itself
fails, then its monitor, if any, gets the notification.

It is possible for one monitor to watch many coroutines. Because of
this, coroutine exceptions are now maintained with a list.

All this is a bit experimental for now