Added features to locate peers and run 'reactors'.
Added facility to register and run remote 'reactors' (for lack of better
term). Reactors are corutines that can be started by remote peers. This
is in lieu of peers sending arbitrary methods to be executed, which is
not only hard to use (although simple to implement in asyncoro, especially
as most of it is done in dispy), but the side effects of importing new
functions can be dangerous. A reactor, on the other hand, is purposely
developed to be called by peers (hopefully it checks against illegal
uses etc.) An alternative would be to use (psuedo)sandbox (e.g., use
subprocess/multiprocessing.Process, as done in dispy), but that would
not be a coroutine!

asyncoro instances can now find each other using name they are registered
with. With this, specific services can be used at a particular asyncoro
(found with 'locate_peer').