Relay pings to other instances on same node.
Earlier implementation assumed multiple servers can bind to same port
and messages would be delivered to all the servers. It doesn't (always)
work (not on Linux, apparently). Now ping messages are relayed to
other asyncoro instances running on same node.

When sending messages in cached peers / requests, make deepcopy so that
interleaving executions don't interfere with one another, as locks are
not used.