Fixed 'deliver' method for remote channels.
'deliver' method should send 'n' (min number of recipients) to
remote asyncoro in the case of remote channel, so the method
blocks until there are 'n' subscribers.

'transform' of Channel, if provided, is executed when sending
messages. This may fail, so protect it with try/except and treat
that case as message to be ignored.