Added asynchronous file and pipe support.
AsyncFile implements asynchronous interface for file-like objects,
for sockets and pipes. Notably, this is not for regular on-disk
files, as they are not "blockable" and pollers don't support event
handilng to indicate read/write events. This class implements
read, write and readline methods.

AsyncPipe implements asynchronous interface for pipes. It
wraps subprocess.Popen objects to provide read, write and
communicate methods. It can be used for a single pipe or connected
pipes.