MonitorException is now sent as a message.
MonitorException is now sent as a message to monitor, instead of
throwing exception. Handling exceptions is a bit cumbersome, whereas
processing messages is much simpler and cleaner. The monitor
can process the messages just as it would any other message; the
type of message can be checked to know if it is MonitorException
or some other message.

Added 'atexit' to AsynCoro so that functions can be registered
that will be executed after the scheduler has terminated.