'coro' default parameter is now optional in Coro generator.
'coro' default/kewyword parameter in the generator methods used
to create coroutines is now optional. If it is present like before
(as 'coro=None'), then the instance of coroutine is set to 'coro'.
If it is not present, it is not an error and if necessary, the
coroutine can obtain it with 'coro = AsynCoro.cur_coro()'.

Made 'cur_coro' a static method in AsynCoro, so coroutines
can obtain instance to running coroutine with it (altough, it
may be more convenient to use 'coro=None' as default parameter in
the generator method instead).

Renamed AsynCoroSocket to AsyncSocket, AsynCoroThreadPool to
AsyncThreadPool, AsynCoroDBCursor to AsyncDBCursor. AsyncThreadPool's
async_task doesn't take 'coro' parameter anymore; it is obtained
with AsynCoro.cur_coro().