Peer status is now sent with message, instead of callback.
Peer status is now sent as message to a registered coroutine.
This is more convenient and flexible, not to mention it is the
way with asyncoro.

Added 'notify' method to coroutine, which is similar to 'monitor',
except it is invoked with arguments reversed (see doc string).

Lot of syntatic changes as required / suggested by 'flake8'.