Added/fixed examples of discoro.
Added more illustrative examples of using discoro for distributed
coroutines.

discoro_client examples in earlier version created the coroutine
to submit jobs in '__main__', but didn't wait for that process
to finish. Because asyncoro waits for non-daemon coroutines to
finish, the examples worked most of the time. However, this is done
with 'atexit' handler, and it seems sometimes by then certain
parts were shutdown by the interpretter (for example status
coroutines wouldn't be added asyncoro scheduler, although they are
created etc.). The fix is to wait for main coroutine to finish,
by calling it's 'value' method.