Don't wrap final value with StopIteration.
When a coroutine is finished, the final value is sent to its
monitors. The value was being wrapped with StopIteration, but the
type of value is already available as first value of tuple, so
no need to do so. Moreover, wrapping with StopIteration was
inconcistent, as when the coroutine stopped with StopIteration
it was sent as the value (without wrapping StopIteration) and when
the last statement is 'yield', it was wrapped, causing the caller
to treat these as different cases.

'sendall' and 'recvall' timeout semantics have changed; earlier,
the timeout was for all given data, whereas now timeout is for
transfer of any data. Using timeout for all data is a good idea
in theory but in practice problematic, as the caller has to use
different timeouts depending on the amount of data to be sent, and
how slow network is.  This patch should've been a separate commit,
but py2 branch already had this committed earlier by mistake.

Also changed _new_generator to _swap_generator to convey the
purpose of the variable more clearly.