Added 'broadcast' option to 'peer' method.
'peer' method in AsynCoro takes 'broadcast' option. This can be
used to find all peers in a remote network - instead of issuing
'peer' to find each peer, 'broadcast' can be used with one peer
in a remote network.

Added 'propagate' to inform asyncoro's running on same node (using
same UDP port). This is automatically implied if 'peer' is used
when host name / IP address is used (i.e., with TCP port 0).

Changed version number to 3.4 as 'peer' method has changed, and
'loc' reference in some places has been changed to 'location' to
be consistent (with the usage elsewhere).