Added 'node' option to add remote nodes.
'node' option can be used as many times as needed to add remote
networks. Only one node is needed to be added per network; discoro
uses 'broadcast' option to discover other nodes in that network.

Transfer files to remote server process even if the process is
running on the same node as the client, so the files would be
in same path as they would be in any other node.