Added zombie detection.
Added 'zombie_period' option to discoro (scheduler). This is the
maximum time a process can stay idle (without executing jobs). If
a process stays idle for the given number of seconds, it closes
the computation. Default zombie_period is None if scheduler is
private and 1800 seconds (30 minutes) if scheduler is shared (i.e.,
run as program). If scheduler is private, user can use the scheduler
and nodes exclusively so there is no reason to terminate computation
automatically. However, if scheduler is shared, other users may
submit computations, so closing idle computations may help other
computations to use resources.