Fixed handshake message proessing.
Peers identify each other with 'ping' and 'pong' messages. When
a peer is identified, any pending asynchronous messages are sent
to that peer. Since each peer broadcasts ping and other peers
respond to that by sending ping themselves, and these messages
are processed asynchronously, handshake may happpen while processing
'ping' or 'pong', so we need to make sure that pending messages are
sent at either, but not both. This was done by checking if the peer
is already in _Peer.peers dictionary before adding (without 'yield')
in between).

If a peer terminates normally, the peer broadcasts that it is
quitting and other peers remove it from _Peer.peers and everything
works as expected. However, if a peer is killed and new peer is
started at the same TCP port where old peer was running, then simply
checking if peer in _Peer.peers is not a valid check, as the old
peer would be still there, but the authentication code for it won't
work with the new peer, preventing peers to communicate. Instead,
peer should be checked for presence as well as authentication match.
Thanks to Oscar Curero for identifying this issue.