Updated documentation in examples.
If in the main thread a coroutine is created as last statement,
as done in many discoro examples, the interpretter may exit (or
start to) before asyncoro's scheduler thread has a chance to
start (which, once started will prevent interpretter to quit). So
'value' method of last created coroutine should be called to allow
asyncoro scheduler to finish the coroutine.