Fixed yet another (last?) issue in clean exit.
When peers are closing, timeout (2 seconds) is used (in case remote
peer is not reachable, local peer at most waits timeout period).
The timeout is not removed in case of successful delivery of
message (it is more efficient to just cleanup while going through
invalid timeouts instead of maintaining and cleaning timeouts).
The fix now is to introduce a dummy short timeout in '_exit' so that
the scheduler doesn't wait for next timeout. This should fix
closing asyncoro when timeouts are used in coroutines that didn't
wait until timeouts (e.g., daemons).

Now discoronode cleanly exits without (sometimes) waiting for timeout.