Fixed scheduler so client doesn't need to wait.
Until now discoro clients would have to call 'value()' on
the client coroutine so that asyncoro/discoro schedulers
are useful; otherwise, scheduling computation would fail.
The real cause for this seemed to be not easy to understand,
so went with 'value()' method approach. However, this is
not a sound / clean approach.

The problem turned out be with discoro scheduler registering
'atexit' to close daemon coroutines (timer, status and
client coroutines) with the idea of cleaning up when
program is exiting. However, when client program's main thread
is done (after creating client coroutine that still can
schedule computation, schedule remote coroutines etc.),
atexit would cause these daemons to terminate, making
scheduler useless. The fix is rather simple: Not register atexit
to terminate these daemons, as asyncoro scheduler will terminate
them anyway when program terminates. If proper cleanup is necessary
for other reasons, asyncoro's "atexit" feature could be used (which
was the reason for adding "atexit" with asyncoro scheduler, but
I myself forgot about it).

The examples are also fixed so 'value()' is not called on client
coroutines anyomre.