Added option 'discover_peers'.
Added option 'discover_peers' to control if discovery message should
be broadcast when asyncoro starts. Default for asyncoro is same as
before: The discovery message is broadcast so all peers are automaitcally
contacted.

With discoronode default is to not discover peers (i.e., it initializes
asyncoro with 'discover_peers=False'). But if option '--discover_peers'
is used on command line, asyncoro is started with 'discover_peers=True'.
In most cases, the servers don't need to communicate with each other;
instead, they only need to communicate with client/scheduler. In such
cases, it is more efficient to disable discovering peers. There is also
'peer' option added to selectively discover given peers; this option can
be used to contact scheduler that is already running, or even to contact
discoro server in remote network.