Send pulse messages from main process.
Start asyncoro in '__main__' thread and send pulse messages to
scheduler from it, instead of from one of the server processes.
With this, the server processes can execute any computation, even
CPU bound computations (that would prevent scheduler from getting
a turn to execute). Coroutines for client are also scheduled only
after successfully sending its info to scheduler (so scheduler
won't give up on it, if computation is CPU bound).