Use auth string to validate server messages.
Timer coroutine in main thread sends pulse messages to current
scheduler, but is not aware of status of servers (e.g., when
they receive computations from scheduler, when a computation is
closed etc.). It is possible for one server to be executing a
computation while another closes and sends 'close' message to
timer (and, it is also possible for this server to accept a new
computation from same / different scheduler). Timer coroutine now
keeps track of auth strings to resolve start / close messages
that may come out of order.