Added 'ping_interval' option to Computation.
By default discoronode servers don't broadcast their address so
if a scheduler is already running, those servers may not be
detected by scheduler (scheduler discovers nodes running when it
starts). 'ping_interval' to Computation can be used to set an
interval when scheduler broadcasts discover messages so any
nodes that were not running when scheduler starts, or any nodes
that were not detected (which may happen if network is noisy, such
as wifi which lose UDP packets used for broadcasting).