Fixed race condition with remote messages.
Messages for remote peers are sent with _Peer, mostly from user
coroutines, but messages are sent with "system" coroutines in
another thread. Although deque itself is thread-safe (at least
the methods used in _Peer), blocks of statements used in checking
if queue is empty or not is not. So use locking as approrpriate,
and message passing to signal queue has items to process (Event
is not appropriate for this).