Fixed storing/retrieving locations and coros.
Locations are sent to web client as strings, but httpd module uses
Location instances, so convert strings back to Location instances
when accessing servers.

Similarly, coro is sent back from client as string and since there
is no way to convert it back to Coro instance, use strings in httpd
module as well.