Use string representations for locations of servers.
The clients use strings for locations / coros. In previous commit,
locations are used as Location instances to be used as keys for
dictionary. This is a bit more expensive operation (to convert
strings back to Location instances) and inconsistent (as coros
are used as strings always). So, use locations as strings for
keys.