Added support for initializing nodes (Posix only).
Added options 'node_available' and 'node_setup' so computations
can send (separate) files to nodes and then initialize them. This
feature can be used for in-memory processing, for example. Memory
initialized thus is visible to all server processes (in earlier
versions memory is initialized in each server process separately,
so if data to be read in memory is same, then there would be
multiple copies of it, whereas now there wouldn't be). However,
this feature works with Posix systems only, as Windows doesn't
support 'fork'.

Changed DiscoroNodeFilter to DiscoroNodeAllocate so node allocation
can be customized (in addition to filtering).