Use nonlocal variable in IOCP's recvall.
In Python 2.7 AllocateREeadBuffer returns a buffer that can be
used to recieve data once (i.e., if read buffer is smaller than
requested, remaining buffer can't be used for next call). So buffer
of at most 1MB is used again and again until requested 'bufsize'
is read. In the last fer iterations of rewriting 'recvall', I/O
completion routine ('_recvall') is made into closure (from partial
function) and then back into partial function with the idea to
improve performance (as closures seem to be slightly better than
partial functions). However, using buffer as argument to '_recvall'
and allocating buffers again and again may result in allocating
too much space, as buffers are not freed until after 'recvall' is
finished. If requested 'bufsize' is large and each WSARecv reads
small fraction of it each time, this can result in allocating too
much space. In this patch nonlocal variable is used to allocate
one buffer at any time. Essentially this is a revert of revert,
although for a different reason than the last revert.