Improved IPv6 support.
With IPv6, interface number is required for UDP broadcast to work.
This is obtained by 'getaddrinfo' with interface name appended, as
'if_nametonumber' is not available until Python 3.5 or so. Interface
address to use is also improved with 'netifaces' and link-local and
unique-local address prefixes.

With these changes IPv6 works in Linux and OS X. I don't have a
Windows box to test, but apparently OS X is more finicky one so
hoping IPv6 works now in Windows as well.