Fixed I/O event processing, especially with SSL.
4.1 version implemented I/O event processing using separate thread
with the idea that messages can be transferred even if user
computations (with discoro, for example) take long time before
'yield'. While this improved usability (programs don't have to
worry about having to 'yield' often if there is no need), it turns
out there are race conditions, especially with SSL (issue raised
in forum at
https://sourceforge.net/p/dispy/discussion/1771151/thread/acbf73a4/).
This patch does away with separate thread for I/O processing, which
is not really needed, as I/O is done with SysCoro that is not meant
for user computations.

Tested with SSL under Linux and OS X.