Fixed issue with initializing nodes.
Node initialization is started in two different paths: One when
a node is discovered right after computation is scheduled and
one where a computation is scheduled right after a node is found.
So there is a chance node initialization is started concurrently,
causing initiliaztion to fail. This patch uses locking to prevent
concurrent node initialization.